change wage and total_cost to decimal in migrations, add total_cost to seeds

+++ server/db/migrations/20230405222114_attendees.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
   return knex.schema.createTable('attendees', (table) => {
     table.increments('id').primary()
     table.string('name')
-    table.integer('wage')
+    table.decimal('wage')
   })
 }
